% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vpro-package}
\alias{vpro}
\alias{vpro-package}
\title{vpro: Manage and Classify Ecological Vegetation Plot Data}
\description{
Provides tools for managing, analysing, and classifying ecological vegetation plot data and hierarchical classifications. VPro (Vegetation Processor) was originally developed as a Microsoft Access application by the British Columbia Ministry of Forests to support the Biogeoclimatic Ecosystem Classification (BEC) program.
}
\author{
\strong{Maintainer}: Will MacKenzie \email{Will.MacKenzie@GEMS1.gov.bc.ca}

Authors:
\itemize{
  \item Will MacKenzie \email{Will.MacKenzie@GEMS1.gov.bc.ca}
  \item Bruno Tremblay \email{bruno@boostao.ca} (\href{https://orcid.org/0000-0002-2945-356X}{ORCID})
  \item Russell Klassen \email{russk@workbc.ca}
}

Other contributors:
\itemize{
  \item  Province of British Columbia [copyright holder, funder]
}

}
\keyword{internal}
