% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{vpro_config_dir}
\alias{vpro_config_dir}
\title{Locate the VPRO user configuration directory}
\usage{
vpro_config_dir()
}
\value{
A normalized directory path.
}
\description{
Locate the VPRO user configuration directory
}
