% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{vpro_config_get}
\alias{vpro_config_get}
\title{Read a VPRO configuration value}
\usage{
vpro_config_get(section, key)
}
\arguments{
\item{section}{Configuration section.}

\item{key}{Configuration key. If omitted, returns the complete section.}
}
\value{
The requested configuration value or section.
}
\description{
Read a VPRO configuration value
}
