% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{vpro_config_set}
\alias{vpro_config_set}
\title{Set a VPRO configuration value}
\usage{
vpro_config_set(section, key, value)
}
\arguments{
\item{section}{Configuration section.}

\item{key}{Configuration key.}

\item{value}{Replacement value.}
}
\value{
\code{value}, invisibly.
}
\description{
Set a VPRO configuration value
}
