% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-connection.R
\name{vpro_db_list}
\alias{vpro_db_list}
\title{List databases attached to a VPRO coordinator}
\usage{
vpro_db_list(con)
}
\arguments{
\item{con}{A VPRO DuckDB connection.}
}
\value{
A character vector of database aliases.
}
\description{
List databases attached to a VPRO coordinator
}
