% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-connection.R
\name{vpro_db_path}
\alias{vpro_db_path}
\title{Construct a path to a VPRO database}
\usage{
vpro_db_path(database, ..., root = vpro_data_dir(), extension = "db")
}
\arguments{
\item{database}{Database name without an extension.}

\item{...}{Optional subdirectories, such as \code{"projects"}.}

\item{root}{VPRO user data directory.}

\item{extension}{Database file extension.}
}
\value{
A character vector of paths.
}
\description{
Construct a path to a VPRO database
}
