% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy-context.R
\name{vpro_hierarchy_deactivate}
\alias{vpro_hierarchy_deactivate}
\title{Deactivate the current VPRO hierarchy}
\usage{
vpro_hierarchy_deactivate(context)
}
\arguments{
\item{context}{A VPRO project context.}
}
\value{
\code{TRUE}, invisibly. Attached hierarchies remain attached.
}
\description{
Deactivate the current VPRO hierarchy
}
