% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-logging.R
\name{vpro_session_logout}
\alias{vpro_session_logout}
\title{Close one VPRO user-login row}
\usage{
vpro_session_logout(context, login, time = Sys.time())
}
\arguments{
\item{context}{A VPRO project context.}

\item{login}{A token returned by \code{vpro_session_login()}.}

\item{time}{Logout time.}
}
\value{
\code{TRUE} if this open session was closed; \code{FALSE} if it was already
closed. An unknown rowid is an error.
}
\description{
Close one VPRO user-login row
}
