% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su-context.R
\name{vpro_su_deactivate}
\alias{vpro_su_deactivate}
\title{Deactivate the current VPRO site-unit table}
\usage{
vpro_su_deactivate(context)
}
\arguments{
\item{context}{A VPRO project context with an active project.}
}
\value{
\code{TRUE}, invisibly. Attached SU tables remain attached.
}
\description{
Deactivate the current VPRO site-unit table
}
