% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whats-new.R
\name{vpro_whats_new_mark_all_viewed}
\alias{vpro_whats_new_mark_all_viewed}
\title{Mark all What's New messages as viewed}
\usage{
vpro_whats_new_mark_all_viewed(context)
}
\arguments{
\item{context}{A VPRO project context with VPro64 attached.}
}
\value{
Number of updated messages, invisibly.
}
\description{
Mark all What's New messages as viewed
}
